example/python-flask/requirements.txt
Flask==0.12.2
python-jose==1.4.0
requests==2.18.4